Refactor: Consolidate Supabase clients and authentication actions
