+++ app/auth/confirm/route.ts
@@ -4,7 +4,7 @@ import type { NextRequest } from "next/server";
 import { redirect } from "next/navigation";
 
 // import { createClient } from "@/lib/supabase/server";
-import { createClient } from "@/utils/supabase/server";
+import { createClient } from "@/lib/supabase/server";
 
 export async function GET(request: NextRequest) {
   const { searchParams } = new URL(request.url);
