+++ lib/offline/mobile-sync-manager.ts
@@ -1,6 +1,6 @@
 import { toast } from "sonner";
 
-import { createClient } from "@/utils/supabase/client";
+import { createClient } from "@/lib/supabase/client";
 
 import { mobileOfflineStorage } from "./mobile-storage";
